+++ src/main/java/com/suyuri/ttodokproject/controller/TextController.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.controller;
+
+public class TextController {
+}
