+++ src/main/java/com/suyuri/ttodokproject/controller/WordController.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.controller;
+
+public class WordController {
+}
