+++ src/main/java/com/suyuri/ttodokproject/controller/quiz/QuizAnswerController.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.controller.quiz;
+
+public class QuizAnswerController {
+}
