+++ src/main/java/com/suyuri/ttodokproject/controller/quiz/QuizChoiceController.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.controller.quiz;
+
+public class QuizChoiceController {
+}
