+++ src/main/java/com/suyuri/ttodokproject/controller/quiz/QuizController.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.controller.quiz;
+
+public class QuizController {
+}
