+++ src/main/java/com/suyuri/ttodokproject/entity/WordEntity.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.entity;
+
+public class WordEntity {
+}
