+++ src/main/java/com/suyuri/ttodokproject/entity/quiz/QuizAnswerEntity.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.entity.quiz;
+
+public class QuizAnswerEntity {
+}
