+++ src/main/java/com/suyuri/ttodokproject/entity/quiz/QuizChoiceEntity.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.entity.quiz;
+
+public class QuizChoiceEntity {
+}
