+++ src/main/java/com/suyuri/ttodokproject/entity/quiz/QuizQuestionEntity.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.entity.quiz;
+
+public class QuizQuestionEntity {
+}
