+++ src/main/java/com/suyuri/ttodokproject/repository/TextRepository.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.repository;
+
+public class TextRepository {
+}
