+++ src/main/java/com/suyuri/ttodokproject/repository/WordRepository.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.repository;
+
+public class WordRepository {
+}
