+++ src/main/java/com/suyuri/ttodokproject/service/WordService.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.service;
+
+public class WordService {
+}
