+++ src/main/java/com/suyuri/ttodokproject/service/quiz/QuizAnswerService.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.service.quiz;
+
+public class QuizAnswerService {
+}
