+++ src/main/java/com/suyuri/ttodokproject/service/quiz/QuizChoiceService.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.service.quiz;
+
+public class QuizChoiceService {
+}
