+++ src/main/java/com/suyuri/ttodokproject/service/quiz/QuizService.java
@@ -1,0 +1,4 @@
+package com.suyuri.ttodokproject.service.quiz;
+
+public class QuizService {
+}
